+++ tests/_mocks/ItemStub.php
@@ -3,40 +3,43 @@
 namespace Swis\JsonApi\Client\Tests\Mocks;
 
 use DateTime;
-use Swis\JsonApi\Client\Model;
+use Swis\JsonApi\Client\Item;
 
-class ModelStub extends Model
+class ItemStub extends Item
 {
     protected $hidden = ['password'];
 
     protected $casts = [
-    'age' => 'integer',
-    'score' => 'float',
-    'data' => 'array',
-    'active' => 'bool',
-    'secret' => 'string',
-    'count' => 'int',
-    'object_data' => 'object',
-    'collection_data' => 'collection',
-    'foo' => 'bar',
-  ];
+        'age' => 'integer',
+        'score' => 'float',
+        'score_inf' => 'float',
+        'score_inf_neg' => 'float',
+        'score_nan' => 'float',
+        'data' => 'array',
+        'active' => 'bool',
+        'secret' => 'string',
+        'count' => 'int',
+        'object_data' => 'object',
+        'collection_data' => 'collection',
+        'foo' => 'bar',
+    ];
 
     protected $guarded = [
-    'secret',
-  ];
+        'secret',
+    ];
 
     protected $fillable = [
-    'name',
-    'city',
-    'age',
-    'score',
-    'data',
-    'active',
-    'count',
-    'object_data',
-    'default',
-    'collection_data',
-  ];
+        'name',
+        'city',
+        'age',
+        'score',
+        'data',
+        'active',
+        'count',
+        'object_data',
+        'default',
+        'collection_data',
+    ];
 
     public function getListItemsAttribute($value)
     {